Module Html_syntax.Attr

type t = Virtual_dom.Vdom.Attr.t

include module type of Attr

type align_options = Attr.align_options =

| None

| X_max_y_max

| X_max_y_mid

| X_max_y_min

| X_mid_y_max

| X_mid_y_mid

| X_mid_y_min

| X_min_y_max

| X_min_y_mid

| X_min_y_min

type units = Attr.units =

| Object_bounding_box

| User_space_on_use

type angle = Attr.angle =

| Deg of float

| Grad of float

| Rad of float

| Turn of float

type path_op = Attr.path_op =

| Move_to_abs of {

x : float;

y : float;

}

| Move_to_rel of {

x : float;

y : float;

}

| Line_to_abs of {

x : float;

y : float;

}

| Line_to_rel of {

x : float;

y : float;

}

| Arc_to_abs of {

rx : float;

ry : float;

x_axis_rotation : float;

large_arc : bool;

sweep : bool;

x : float;

y : float;

}

| Arc_to_rel of {

rx : float;

ry : float;

x_axis_rotation : float;

large_arc : bool;

sweep : bool;

dx : float;

dy : float;

}

| Cubic_abs of {

x1 : float;

y1 : float;

x2 : float;

y2 : float;

x : float;

y : float;

}

| Cubic_rel of {

x1 : float;

y1 : float;

x2 : float;

y2 : float;

x : float;

y : float;

}

| Cubic_smooth_abs of {

x2 : float;

y2 : float;

x : float;

y : float;

}

| Cubic_smooth_rel of {

x2 : float;

y2 : float;

x : float;

y : float;

}

| Quadratic_abs of {

x1 : float;

y1 : float;

x : float;

y : float;

}

| Quadratic_rel of {

x1 : float;

y1 : float;

x : float;

y : float;

}

| Quadratic_smooth_abs of {

x : float;

y : float;

}

| Quadratic_smooth_rel of {

x : float;

y : float;

}

| Close_path

type transform_op = Attr.transform_op =

| Matrix of float * float * float * float * float * float

| Translate of {

dx : float;

dy : float;

}

| Scale of {

sx : float;

sy : float;

}

| Rotate of {

a : [ `Deg of float ];

x : float;

y : float;

}

| Skew_x of float

| Skew_y of float

type text_anchor = Attr.text_anchor =

| Start

| Middle

| End

val viewbox : min_x:float -> min_y:float -> width:float -> height:float -> Virtual_dom.Vdom.Attr.t

<a>

val href : string -> Virtual_dom.Vdom.Attr.t

many things

val id : here:lexing_position -> string -> Virtual_dom.Vdom.Attr.t

val transform : transform_op list -> Virtual_dom.Vdom.Attr.t

<circle>

val cx : float -> Virtual_dom.Vdom.Attr.t

val cy : float -> Virtual_dom.Vdom.Attr.t

val r : float -> Virtual_dom.Vdom.Attr.t

nothing for <defs>

<ellipse>

shares cx and cy with circle

val rx : float -> Virtual_dom.Vdom.Attr.t

val ry : float -> Virtual_dom.Vdom.Attr.t

nothing for <g>

<image>

val x : float -> Virtual_dom.Vdom.Attr.t

val y : float -> Virtual_dom.Vdom.Attr.t

val width : float -> Virtual_dom.Vdom.Attr.t

val height : float -> Virtual_dom.Vdom.Attr.t

val xlink_href : string -> Virtual_dom.Vdom.Attr.t

val preserve_aspect_ratio : align:align_options -> ?meet_or_slice:[ `Meet | `Slice ] -> unit -> Virtual_dom.Vdom.Attr.t

<line>

val x1 : float -> Virtual_dom.Vdom.Attr.t

val x2 : float -> Virtual_dom.Vdom.Attr.t

val y1 : float -> Virtual_dom.Vdom.Attr.t

val y2 : float -> Virtual_dom.Vdom.Attr.t

<linearGradient>

shares href with <a>, x1, x2, y1, y2 with <line>

val gradient_units : units -> Virtual_dom.Vdom.Attr.t

val gradient_transform : transform_op list -> Virtual_dom.Vdom.Attr.t

<marker>

viewBox is shared with <svg>, preserveAspecteRatio is shared with <image>

val marker_height : float -> Virtual_dom.Vdom.Attr.t

val marker_width : float -> Virtual_dom.Vdom.Attr.t

val marker_units : units -> Virtual_dom.Vdom.Attr.t

val orient : [ `Angle of angle | `Auto | `Auto_start_reverse ] -> Virtual_dom.Vdom.Attr.t

val refX : float -> Virtual_dom.Vdom.Attr.t

val refY : float -> Virtual_dom.Vdom.Attr.t

<mask>

shares width, height, x, and y with <image>

val mask_units : units -> Virtual_dom.Vdom.Attr.t

val mask_content_units : units -> Virtual_dom.Vdom.Attr.t

<path>

val d : path_op list -> Virtual_dom.Vdom.Attr.t

<polygon>

val points : (float * float) list -> Virtual_dom.Vdom.Attr.t

val fill : [< `Url of string | Css_gen.Color.t ] -> Virtual_dom.Vdom.Attr.t

val stroke : [< Css_gen.Color.t ] -> Virtual_dom.Vdom.Attr.t

val stroke_width : float -> Virtual_dom.Vdom.Attr.t

val stroke_linecap : [ `Butt | `Round | `Square ] -> Virtual_dom.Vdom.Attr.t

val stroke_dasharray : float list -> Virtual_dom.Vdom.Attr.t

val stroke_dashoffset : float -> Virtual_dom.Vdom.Attr.t

<polyline>

shares points with <polygon>

<radialGradient>

shares spreadMethod with <linearGradient> shares cx, cy, r with <circle>

val fx : float -> Virtual_dom.Vdom.Attr.t

val fy : float -> Virtual_dom.Vdom.Attr.t

val fr : float -> Virtual_dom.Vdom.Attr.t

val spread_method : [ `Pad | `Reflect | `Repeat ] -> Virtual_dom.Vdom.Attr.t

Nothing to do for <rect> because shares x, y, rx and ry width, height with <image>

<stop>

val offset : Core.Percent.t -> Virtual_dom.Vdom.Attr.t

val stop_color : [< Css_gen.Color.t ] -> Virtual_dom.Vdom.Attr.t

val stop_opacity : Core.Percent.t -> Virtual_dom.Vdom.Attr.t

nothing to do for <style>

Nothing to do for <symbol> as refX and refY is shared with <marker>, viewBox is shared with <svg>, x, y, width, and height are shared with <image>,

<text>

x and y is shared with <image>

val dx : float -> Virtual_dom.Vdom.Attr.t

val dy : float -> Virtual_dom.Vdom.Attr.t

module Text = Attr.Text

val font_size : here:lexing_position -> Css_gen.Length.t -> Virtual_dom.Vdom.Attr.t

val text_anchor : here:lexing_position -> text_anchor -> Virtual_dom.Vdom.Attr.t

val (@) : Virtual_dom.Vdom.Attr.t -> Virtual_dom.Vdom.Attr.t -> Virtual_dom.Vdom.Attr.t

module Primitives = Virtual_dom.Vdom.Html_syntax.Html_syntax.Attr.Primitives